[refactor/#5] Home뷰 작품 추천리스트 컴포넌트
